+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### New
 
 ### Changes
+* Added a CC Email to payment emails
 * Added Receiver Name to Banking Information
 
 ### Fixes